mods_panel create sorting.py and move class ModsPanelSortKey to it
