+++ app/views/mods_panel.py
@@ -1,6 +1,5 @@
 import json
 import os
-from enum import Enum
 from functools import partial
 from pathlib import Path
 from shutil import copy2, copytree
@@ -53,6 +52,7 @@ from app.utils.generic import (
     sanitize_filename,
 )
 from app.utils.metadata import MetadataManager, ModMetadata
+from app.utils.sorting import ModsPanelSortKey, sort_uuids
 from app.views.deletion_menu import ModDeletionMenu
 from app.views.dialogue import (
     show_dialogue_conditional,
@@ -61,54 +61,6 @@ from app.views.dialogue import (
 )
 from app.views.mods_panel_icons import ModListIcons
 from app.views.mods_panel_list_inner import ModListItemInner
-
-
-def uuid_no_key(uuid: str) -> str:
-    """
-    Returns the UUID of the mod.
-    Args:
-        uuid (str): The UUID of the mod.
-    Returns:
-        str: The UUID of the mod.
-    """
-    return uuid
-
-
-def uuid_to_mod_name(uuid: str) -> str:
-    """
-    Converts a UUID to the corresponding mod name.
-    Args:
-        uuid (str): The UUID of the mod.
-    Returns:
-        str: If mod name not None, returns mod name in lowercase. Otherwise, returns "# unnamed mod".
-    """
-    name = MetadataManager.instance().internal_local_metadata[uuid]["name"]
-    return name.lower() if name is not None else "# unnamed mod"
-
-
-class ModsPanelSortKey(Enum):
-    """
-    Enum class representing different sorting keys for mods.
-    """
-
-    NOKEY = 0
-    MODNAME = 1
-
-
-def sort_uuids(uuids: list[str], key: ModsPanelSortKey) -> list[str]:
-    """
-    Sort the list of UUIDs based on the provided key.
-    Args:
-        key (ModsPanelSortKey): The key to sort the list by.
-    Returns:
-        None
-    """
-    # Sort the list of UUIDs based on the provided key
-    if key == ModsPanelSortKey.MODNAME:
-        key_function = uuid_to_mod_name
-    else:
-        return sorted(uuids, key=lambda x: x)
-    return sorted(uuids, key=key_function)
 
 
 class ModListWidget(QListWidget):
